Add files via upload
For some reason my github was spazzing out on my linux box. Could not commit normally (and failed to commit the day that I initially tried without me realizing), so I attempted to upload all the files via the website. The website wouldn't take more than 100 files, so I zipped it and uploaded it. Sorry for the lazyness.